requirements.txt
tbapy
pyyaml
flask
flask-socketio
